+++ lib/tracker_logger_bundle.dart
@@ -80,8 +80,8 @@ class TrackerLoggerBundle extends Plugin {
   @override
   void logEvent(
     String name, {
-    String? category,
-    String? action,
+    required String category,
+    required String action,
     int? value,
     LogLevel level = LogLevel.info,
   }) {