+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 
+## 0.0.3
+
+- Fix error logging with hint
+- Make action required for event logging
+
 ## 0.0.2
 
-* Improve pub points by adding an example, improving the readme and fixing the SDK version
+- Improve pub points by adding an example, improving the readme and fixing the SDK version
 
 ## 0.0.1
 
-* Set up the package.
+- Set up the package.